interesting commit
I don't know why, but my app's transitioned state for the menu-div-container was in the middle of the page.  But oddly enough, the glitchy delays in the transition of the menu style had disappeared.

I honestly have no idea why it's suddenly performing better and why the transitioned state of the menu had changed.  it was a quick fix though.
I'll test it in the github.io page as well to see if the performance delays are still there.  If so, I could try to learn about what aspects of storage or caching or loading only necessary materials improves the performance.